+++ .flake8
@@ -1,3 +1,4 @@
 [flake8]
+ignore = E203 W503
 max-line-length = 120
 exclude = .tox,.git,*/migrations/*,*/static/CACHE/*,docs,node_modules
