+++ pyproject.toml
@@ -57,7 +57,7 @@ dependencies = [
     # https://github.com/encode/django-rest-framework
 
     # OpenApi Support
-    'drf-spectacular==0.27.2',
+    'drf-spectacular>=0.27.2',
 
     # tagging support
     'django-tagulous>=1.3.0',
