+++ topobank/manager/containers.py
@@ -1,6 +1,7 @@
 """
 Import and export surfaces through archives aka "surface containers".
 """
+
 import datetime
 import json
 import logging
@@ -24,20 +25,20 @@ from .models import Property, Surface, Topography, topography_datafile_path
 _log = logging.getLogger(__name__)
 
 TOPOGRAPHY_DEFAULT_ATTR_VALUES = {
-    'tags': [],
-    'fill_undefined_data_mode': Topography.FILL_UNDEFINED_DATA_MODE_NOFILLING,
-    'detrend_mode': 'center',
-    'is_periodic': False,
-    'height_scale': 1.,
-    'data_source': 0,
-    'measurement_date': datetime.date(1970, 1, 1),
-    'description': '',
+    "tags": [],
+    "fill_undefined_data_mode": Topography.FILL_UNDEFINED_DATA_MODE_NOFILLING,
+    "detrend_mode": "center",
+    "is_periodic": False,
+    "height_scale": 1.0,
+    "data_source": 0,
+    "measurement_date": datetime.date(1970, 1, 1),
+    "description": "",
 }
 
 SURFACE_DEFAULT_ATTR_VALUES = {
-    'tags': [],
-    'description': '',
-    'is_published': False,
+    "tags": [],
+    "description": "",
+    "is_published": False,
 }
 
 
@@ -57,9 +58,9 @@ def import_topography(topo_dict, topo_file, surface, ignore_missing=False):
         If True, try to find reasonable defaults for missing attributes.
         (Default: False)
     """
-    topo_name = topo_dict['name']
+    topo_name = topo_dict["name"]
 
-    size_x, *size_rest = topo_dict['size']
+    size_x, *size_rest = topo_dict["size"]
     size_y = None if len(size_rest) == 0 else size_rest[0]
 
     user = surface.creator
@@ -74,34 +75,36 @@ def import_topography(topo_dict, topo_file, surface, ignore_missing=False):
         surface=surface,
         size_x=size_x,
         size_y=size_y,
-        measurement_date=topo_dict['measurement_date'],
-        description=topo_dict['description'],
-        data_source=topo_dict['data_source'],
-        unit=topo_dict['unit'],
-        tags=topo_dict['tags'],
-        fill_undefined_data_mode=topo_dict['fill_undefined_data_mode'],
-        detrend_mode=topo_dict['detrend_mode'],
-        is_periodic=topo_dict['is_periodic'],
+        measurement_date=topo_dict["measurement_date"],
+        description=topo_dict["description"],
+        data_source=topo_dict["data_source"],
+        unit=topo_dict["unit"],
+        tags=topo_dict["tags"],
+        fill_undefined_data_mode=topo_dict["fill_undefined_data_mode"],
+        detrend_mode=topo_dict["detrend_mode"],
+        is_periodic=topo_dict["is_periodic"],
     )
 
     try:
-        topo_kwargs.update(dict(
-            instrument_name=topo_dict['instrument']['name'],
-            instrument_type=topo_dict['instrument']['type'],
-            instrument_parameters=topo_dict['instrument']['parameters'],
-        ))
+        topo_kwargs.update(
+            dict(
+                instrument_name=topo_dict["instrument"]["name"],
+                instrument_type=topo_dict["instrument"]["type"],
+                instrument_parameters=topo_dict["instrument"]["parameters"],
+            )
+        )
     except KeyError:
         # Metadata does not contain instrument information
         pass
 
     try:
-        topo_kwargs['height_scale'] = topo_dict['height_scale']
+        topo_kwargs["height_scale"] = topo_dict["height_scale"]
     except KeyError:
         # If height_scale is not included, it will probably already
         # applied because of file contents while loading
         pass
 
-    topo_kwargs['datafile'] = topo_name
+    topo_kwargs["datafile"] = topo_name
     topography = Topography(**topo_kwargs)
 
     # We need to save the topography to get an id...
@@ -114,8 +117,14 @@ def import_topography(topo_dict, topo_file, surface, ignore_missing=False):
     topography.save()
 
 
-def import_container(surface_zip, user, datafile_attribute='datafile', ignore_missing=False, unsafe_yaml_loader=False,
-                     tag=None):
+def import_container(
+    surface_zip,
+    user,
+    datafile_attribute="datafile",
+    ignore_missing=False,
+    unsafe_yaml_loader=False,
+    tag=None,
+):
     """
     Import surfaces from a ZIP archive and create corresponding Surface instances in the database.
 
@@ -153,44 +162,52 @@ def import_container(surface_zip, user, datafile_attribute='datafile', ignore_mi
     extensive validation of the archive's contents.
     """
     surfaces = []
-    with surface_zip.open('meta.yml', mode='r') as meta_file:
+    with surface_zip.open("meta.yml", mode="r") as meta_file:
         yaml_loader = yaml.full_load if unsafe_yaml_loader else yaml.safe_load
         meta = yaml_loader(meta_file)
 
-        for surface_dict in meta['surfaces']:
+        for surface_dict in meta["surfaces"]:
             if ignore_missing:
                 for k, default in SURFACE_DEFAULT_ATTR_VALUES.items():
                     surface_dict.setdefault(k, default)
 
             import_time = str(now())
 
-            surface_description = surface_dict['description']
-            surface_description += f'\n\nImported from file "{surface_zip.filename}" on {import_time}.'
-            tags = set(surface_dict['tags'])
+            surface_description = surface_dict["description"]
+            surface_description += (
+                f'\n\nImported from file "{surface_zip.filename}" on {import_time}.'
+            )
+            tags = set(surface_dict["tags"])
             if tag is not None:
                 tags.add(tag)
-            surface = Surface(creator=user,
-                              name=surface_dict['name'],
-                              category=surface_dict['category'],
-                              description=surface_description,
-                              tags=tags)
+            surface = Surface(
+                creator=user,
+                name=surface_dict["name"],
+                category=surface_dict["category"],
+                description=surface_description,
+                tags=tags,
+            )
             surface.save()
 
-            if 'properties' in surface_dict:
-                for property_dict in surface_dict['properties']:
-                    name = property_dict['name']
-                    value = property_dict['value']
-                    unit = property_dict.get('unit', None)
+            if "properties" in surface_dict:
+                for property_dict in surface_dict["properties"]:
+                    name = property_dict["name"]
+                    value = property_dict["value"]
+                    unit = property_dict.get("unit", None)
                     surface.properties.add(Property(name=name, value=value, unit=unit))
                 surface.save()
 
             surfaces += [surface]
 
-            for topo_idx, topo_dict in enumerate(surface_dict['topographies']):
-                datafile_name = topo_dict[datafile_attribute]['original']  # we just import the original
-                topo_file = surface_zip.open(datafile_name, mode='r')
+            for topo_idx, topo_dict in enumerate(surface_dict["topographies"]):
+                datafile_name = topo_dict[datafile_attribute][
+                    "original"
+                ]  # we just import the original
+                topo_file = surface_zip.open(datafile_name, mode="r")
 
-                import_topography(topo_dict, topo_file, surface, ignore_missing=ignore_missing)
+                import_topography(
+                    topo_dict, topo_file, surface, ignore_missing=ignore_missing
+                )
 
     return surfaces
 
@@ -212,9 +229,11 @@ def write_surface_container(file, surfaces):
     """
     surfaces_dicts = []
 
-    publications = set()  # collect publications so we can list the licenses in an extra file
+    publications = (
+        set()
+    )  # collect publications so we can list the licenses in an extra file
 
-    zf = zipfile.ZipFile(file, mode='w')
+    zf = zipfile.ZipFile(file, mode="w")
 
     #
     # Add meta data and topography files for all given surfaces
@@ -222,7 +241,11 @@ def write_surface_container(file, surfaces):
     nb_surfaces = len(surfaces)
     log10_nb_surfaces = int(math.log10(nb_surfaces)) if nb_surfaces > 0 else 0
     for surface_index, surface in enumerate(surfaces):
-        surface_prefix = '' if nb_surfaces == 1 else f'{surface_index}'.zfill(log10_nb_surfaces + 1) + '-'
+        surface_prefix = (
+            ""
+            if nb_surfaces == 1
+            else f"{surface_index}".zfill(log10_nb_surfaces + 1) + "-"
+        )
 
         topographies = Topography.objects.filter(surface=surface)
 
@@ -232,9 +255,13 @@ def write_surface_container(file, surfaces):
         # using the original file name + a counter, if needed
 
         nb_topographies = len(topographies)
-        log10_nb_topographies = int(math.log10(nb_topographies)) if nb_topographies > 0 else 0
+        log10_nb_topographies = (
+            int(math.log10(nb_topographies)) if nb_topographies > 0 else 0
+        )
         for topography_index, topography in enumerate(topographies):
-            topography_prefix = f'{topography_index}'.zfill(log10_nb_topographies + 1) + '-'
+            topography_prefix = (
+                f"{topography_index}".zfill(log10_nb_topographies + 1) + "-"
+            )
 
             topo_dict = topography.to_dict()
             # this dict may be okay, but have to check whether the filename is unique
@@ -245,8 +272,8 @@ def write_surface_container(file, surfaces):
             #
 
             # Split out original extension
-            _, original_extension = os.path.splitext(topo_dict['datafile']['original'])
-            slugified_name = slugify(topo_dict['name'])
+            _, original_extension = os.path.splitext(topo_dict["datafile"]["original"])
+            slugified_name = slugify(topo_dict["name"])
             # `slugified_name` may have an extension since initial `name` is filename
             name, slugified_extension = os.path.splitext(slugified_name)
             if slugified_extension == original_extension:
@@ -256,8 +283,8 @@ def write_surface_container(file, surfaces):
                 # Apparently the `name` was not a filename and we need to add the extension
                 slugified_extension = original_extension
             # Construct filename for use within the container (note: extension contains the leading '.')
-            name_in_container = f'{surface_prefix}{topography_prefix}{slugified_name}{slugified_extension}'
-            topo_dict['datafile']['original'] = name_in_container
+            name_in_container = f"{surface_prefix}{topography_prefix}{slugified_name}{slugified_extension}"
+            topo_dict["datafile"]["original"] = name_in_container
 
             # add topography file to ZIP archive
             zf.writestr(name_in_container, topography.datafile.read())
@@ -269,19 +296,25 @@ def write_surface_container(file, surfaces):
                 try:
                     topography.renew_squeezed_datafile()
                 except Exception as exc:
-                    _log.error(f"Cannot generate squeezed datafile of topography id {topography.id} "
-                               f"for download: {exc}")
+                    _log.error(
+                        f"Cannot generate squeezed datafile of topography id {topography.id} "
+                        f"for download: {exc}"
+                    )
             if topography.has_squeezed_datafile:
-                squeezed_name_in_container = f'{surface_prefix}{topography_prefix}{slugified_name}-squeezed.nc'
-                topo_dict['datafile']['squeezed-netcdf'] = squeezed_name_in_container
+                squeezed_name_in_container = (
+                    f"{surface_prefix}{topography_prefix}{slugified_name}-squeezed.nc"
+                )
+                topo_dict["datafile"]["squeezed-netcdf"] = squeezed_name_in_container
 
                 # add topography file to ZIP archive
-                zf.writestr(squeezed_name_in_container, topography.squeezed_datafile.read())
+                zf.writestr(
+                    squeezed_name_in_container, topography.squeezed_datafile.read()
+                )
 
             topography_dicts.append(topo_dict)
 
         surface_dict = surface.to_dict()
-        surface_dict['topographies'] = topography_dicts
+        surface_dict["topographies"] = topography_dicts
 
         surfaces_dicts.append(surface_dict)
 
@@ -302,7 +335,8 @@ def write_surface_container(file, surfaces):
     #
     # Add a Readme file and license files
     #
-    readme_txt = textwrap.dedent(f"""
+    readme_txt = textwrap.dedent(
+        f"""
     Contents of this ZIP archive
     ============================
     This archive contains {len(surfaces)} digital surface twin(s). Each digital surface
@@ -326,7 +360,8 @@ def write_surface_container(file, surfaces):
 
     TopoBank: {topobank.__version__}
     SurfaceTopography: {SurfaceTopography.__version__}
-    """)
+    """
+    )
 
     if len(publications) > 0:
         #
@@ -334,14 +369,18 @@ def write_surface_container(file, surfaces):
         #
         for pub in publications:
             if pub.doi_name:
-                zf.writestr(f"other/datacite-{pub.short_url}.json", json.dumps(pub.datacite_json))
+                zf.writestr(
+                    f"other/datacite-{pub.short_url}.json",
+                    json.dumps(pub.datacite_json),
+                )
 
         #
         # Add license information to README
         #
         licenses_used = set(pub.license for pub in publications)
         legalcode_filepath = {pub.license: pub.get_license_legalcode_filepath()}
-        readme_txt += textwrap.dedent("""
+        readme_txt += textwrap.dedent(
+            """
         License information
         ===================
 
@@ -349,23 +388,28 @@ def write_surface_container(file, surfaces):
         licenses, please look at the metadata for each surface
         for the specific license:
 
-        """)
+        """
+        )
 
         for license in licenses_used:
             license_file_in_archive = f"LICENSE-{license}.txt"
             license_info = settings.CC_LICENSE_INFOS[license]
-            readme_txt += textwrap.dedent("""
+            readme_txt += textwrap.dedent(
+                """
             {}
             {}
             For details about this license see
             - '{}' (description), or
             - '{}' (legal code), or
             - the included file '{}' (legal code).
-            """.format(license_info['title'],
-                       "-" * len(license_info['title']),
-                       license_info['description_url'],
-                       license_info['legal_code_url'],
-                       license_file_in_archive))
+            """.format(
+                    license_info["title"],
+                    "-" * len(license_info["title"]),
+                    license_info["description_url"],
+                    license_info["legal_code_url"],
+                    license_file_in_archive,
+                )
+            )
             #
             # Also add license file
             #
